cicd/jenkins-pipeline/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-jenkins==1.8.3
pydantic==2.11.3
starlette==0.46.1
